requirements.txt
requests>=2.32.0
python-dotenv>=1.0.1
google-generativeai>=0.8.0
openai>=1.45.0
